backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
asyncpg==0.29.0
pydantic-settings==2.1.0
python-multipart==0.0.6
torch==2.2.0
pandas==2.1.4
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0
pytest-cov==4.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
aiosqlite==0.19.0
jinja2==3.1.3
email-validator==2.1.0
